refactor: move helper fncs to own modules
