+++ dockubeadt/utils.py
@@ -1,3 +1,6 @@
+import re
+import subprocess
+
 from ruamel.yaml import YAML
 
 yaml = YAML()
@@ -35,3 +38,30 @@ def dump_yaml(data, stream):
         stream (file): The stream to dump the YAML data to.
     """
     yaml.dump(data, stream)
+
+def run_command(cmd):
+    """
+    Executes a shell command and returns the output and return code.
+
+    Args:
+        cmd (str): The command to execute.
+
+    Returns:
+        tuple: A tuple containing the return code and output of the command.
+    """
+    with subprocess.Popen(
+            cmd, 
+            stderr=subprocess.STDOUT,
+            stdout=subprocess.PIPE,
+            shell=True
+    ) as p:
+        
+        output = ""
+        for line in p.stdout:
+            # Regex gets rid of additional characters in Kompose output
+            output += re.sub(
+                r'\x1b(\[.*?[@-~]|\].*?(\x07|\x1b\\))',
+                '',
+                line.decode()
+            )
+    return p.returncode, output
